pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "harness"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40",
    "pydantic>=2.0",
    "pandas>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]
